antiplag/antiplag/serializers.py
from rest_framework import serializers
from .models import Paper

class PaperSerializer(serializers.ModelSerializer):
    class Meta:
        model = Paper
        fields = ['title','text','file','timestamp','user']